removed transition, resizing the header on panel toggle
if the header is sticky, when we open or close the side panels, the sticky header doesn't resize itself. if there is a resize animation, I have to wait till it finishes, so I can calculate the new width for the sticky header. But that looks ugly, so I removed complitely the panel sliding animations.